+++ packages/dev/src/pages/collections/[handle].server.jsx
@@ -4,7 +4,7 @@ import {
   useShopQuery,
   flattenConnection,
   RawHtml,
-  ProductSort,
+  ProductSortOption,
 } from '@shopify/hydrogen';
 import {useParams, useLocation} from 'react-router-dom';
 import gql from 'graphql-tag';
@@ -45,13 +45,31 @@ export default function Collection({
   const products = flattenConnection(collection.products);
   const hasNextPage = data.collection.products.pageInfo.hasNextPage;
 
+  const sortOptions = [
+    {title: 'created', sortKey: 'created'},
+    {title: 'price', sortKey: 'price'},
+    {title: 'best selling', sortKey: 'best_selling'},
+  ];
+
   return (
     <Layout>
       <h1 className="font-bold text-4xl md:text-5xl text-gray-900 mb-6 mt-6">
         {collection.title}
       </h1>
       <RawHtml string={collection.descriptionHtml} className="text-2xl" />
-      <ProductSort as="ul" className="border-black	border-2" />
+
+      <ul>
+        {sortOptions.map((sortOption) => (
+          <ProductSortOption
+            key={sortOption.sortKey}
+            sortOption={sortOption}
+            className="border-2 border-red-700 block"
+          >
+            <li>{sortOption.title}</li>
+          </ProductSortOption>
+        ))}
+      </ul>
+
       <p className="text-sm text-gray-500 mt-5 mb-5">
         {products.length} {products.length > 1 ? 'products' : 'product'}
       </p>
