--- packages/hydrogen/src/components/ProductSort/index.ts
@@ -1,1 +1,0 @@
-export {ProductSort} from './ProductSort.client';
