+++ packages/hydrogen/src/components/ProductSortOption/index.ts
@@ -1,0 +1,1 @@
+export {ProductSortOption} from './ProductSortOption.client';
