+++ packages/hydrogen/src/components/index.ts
@@ -57,7 +57,7 @@ export {
   ProductProviderFragment,
 } from './ProductProvider';
 export {ProductDescription} from './ProductDescription';
-export {ProductSort} from './ProductSort';
+export {ProductSortOption} from './ProductSortOption';
 export {ProductTitle} from './ProductTitle';
 export {SelectedVariantPrice} from './SelectedVariantPrice';
 export {SelectedVariantAddToCartButton} from './SelectedVariantAddToCartButton';
